data has been loaded into detail page

+++ lib/model/RecipeInformation.dart
@@ -13,8 +13,8 @@ class RecipeInformation {
   String gaps;
   bool lowFodmap;
   int aggregateLikes;
-  int spoonacularScore;
-  int healthScore;
+  double spoonacularScore;
+  double healthScore;
   String creditsText;
   String license;
   String sourceName;
